# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AuthResponse:
    properties:
      email:
        example: analyst@soc.local
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      error:
        example: Invalid request payload
        type: string
      message:
        example: Detailed error description
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        example: analyst@soc.local
        type: string
      password:
        example: Secret123!
        type: string
    required:
    - email
    - password
    type: object
  dto.MessageResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  dto.RegisterRequest:
    properties:
      email:
        example: analyst@soc.local
        type: string
      password:
        example: Secret123!
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    name: SOC Trainer Development Team
  description: Automated Data-Driven Cyberattack Simulation for Blue Team Training
    platform REST API.
  license:
    name: MIT
  termsOfService: https://github.com/balliolon2/Automated_Data-Driven_Cyberattack_Simulation_for_Blue_Team_Training
  title: SOC Trainer REST API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a signed JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new learner account with email and password
      parameters:
      - description: User Registration Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Register a new learner
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOi..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"